Add Basic Triangle Half Test (#656)
This test is mainly to test that half types in function signature works as expected. Not enabled by default yet.

+++ tools/clang/unittests/HLSL/ShaderOpTest.cpp
@@ -1932,7 +1932,7 @@ void ParseDataFromText(LPCWSTR pText, LPCWSTR pEnd, DXIL::ComponentType compType
     if (compType == DXIL::ComponentType::F16) {
       uint16_t fp16Val = ConvertFloat32ToFloat16(fVal);
       pB = (BYTE *)&fp16Val;
-      V.insert(V.end(), pB, pB + sizeof(float));
+      V.insert(V.end(), pB, pB + sizeof(uint16_t));
     }
     else {
       pB = (BYTE *)&fVal;
